requirements.txt
fastapi
pydantic
psycopg2
PyJWT